Update settings.ts
Usage of `split` in javascript results in `=` characters being stripped from values provided in the cli which is against the intent since all that is provided here is the key/value with the first equals as the split. Change to strict splitting of the string using `substring` and finding the first and only the first `=`.